Simplify condition to conditionFunc
