+++ test/tenant_conditions.go
@@ -12,16 +12,16 @@ import (
 type TenantCheck func(*api.Tenant) bool
 
 // CreateTenantCondition creates a WaitCondition for a tenant with a dedicated check
-func CreateTenantCondition(tenant *api.Tenant, check TenantCheck) WaitCondition {
+func CreateTenantCondition(tenant *api.Tenant, check TenantCheck) WaitConditionFunc {
 	key := fmt.Sprintf("%s/%s", tenant.GetNamespace(), tenant.GetName())
-	return NewWaitCondition(func(ctx context.Context) (bool, error) {
+	return func(ctx context.Context) (bool, error) {
 		fetcher := k8s.NewTenantFetcher(GetClientFactory(ctx))
 		tenant, err := fetcher.ByKey(key)
 		if err != nil {
 			return true, err
 		}
 		return check(tenant), nil
-	})
+	}
 }
 
 // TenantHasStateResult creates a TenantCheck which checks for a dedicated State
